Para vista_alertas#casa

+++ resources/views/contract/contracts/show.blade.php
@@ -73,13 +73,13 @@ p.centrado {
                                             <h5>Llamado: {{ $contract->description." - ".$contract->modality->description." N° ".$contract->number_year." - ".$contract->provider->description }}</h5>
                                         </div>
                                         <div class="col-sm-10 text-left">
-                                            <h5 style="font-size: 17px; font-weight: bold; color:#FF0000">Dependencia Responsable: {{ $contract->dependency->description }}</h5>
+                                            <h5 style="font-size: 17px; font-weight: bold; color:blue">Dependencia Responsable: {{ $contract->dependency->description }}</h5>
                                         </div>
                                         <div class="col-sm-2">
-                                                @if (Auth::user()->hasPermission(['admin.orders.create']))
-                                                    @if (in_array($contract->contract_state_id, [1,2]))
+                                                @if (Auth::user()->hasPermission(['admin.orders.create','admin.orders.update']))
+                                                    {{-- @if (in_array($contract->contract_state_id, [1,2])) --}}
                                                         <button class="btn btn-primary dropdown-toggle waves-effect" type="button" id="acciones" data-toggle="dropdown" aria-haspopup="true" aria-expanded="true">Acciones</button>
-                                                    @endif
+                                                    {{-- @endif --}}
                                                 @endif
 
                                                 <div class="dropdown-menu" aria-labelledby="acciones" data-dropdown-in="fadeIn" data-dropdown-out="fadeOut">
@@ -126,7 +126,7 @@ p.centrado {
 
                                     <div class="tab-content card-block">
                                         <div class="tab-pane active" id="tab1" role="tabpanel">
-                                            <h5 class="text-center">Datos del LLAMADO</h5>
+                                            <h5 class="text-center">Datos del LLamado</h5>
                                             <table class="table table-striped table-bcontracted">
                                                 <tbody>
                                                     <tr>
@@ -160,7 +160,7 @@ p.centrado {
                                                         <td>{{ $contract->provider->description }}</td>
 
                                                         {{-- SI ESTADO ESTA ANULADO, RESCINDIDO O CANCELADO --}}
-                                                        @if (in_array($contract->contract_state_id, [2,3,4]))
+                                                        @if (in_array($contract->contract_state_id, [2,3,4,6]))
                                                             <td style="color:#ff0000">{{ $contract->contractState->description }}</td>
                                                         @else
                                                             <td style="color:green">{{ $contract->contractState->description }}</td>
@@ -219,18 +219,22 @@ p.centrado {
                                                                 <td>{{ $contract->items[$i]->AmountFormat()}} </td>
                                                                 <td>{{ $contract->items[$i]->comments }}</td>
                                                                 <td>
-                                                                {{-- @if (Auth::user()->hasPermission(['admin.items.update','contracts.items.update']) || $contract->dependency_id == Auth::user()->dependency_id) --}}
-                                                                @if (Auth::user()->hasPermission(['admin.items.update','contracts.items.update']))
-                                                                    <button type="button" title="Editar" class="btn btn-warning btn-icon" onclick="updateItem({{ $contract->items[$i]->id }})">
-                                                                        <i class="fa fa-pencil"></i>
-                                                                    </button>
+                                                                
+                                                                {{-- No muestra si estado de llamado no es rescindido, cerrado, impugando o en proceso de rescisión --}}
+                                                                @if (in_array($contract->contract_state_id, [1,5]))                                                                
+                                                                    @if (Auth::user()->hasPermission(['admin.items.update','contracts.items.update']))                                                                
+                                                                        <button type="button" title="Editar" class="btn btn-warning btn-icon" onclick="updateItem({{ $contract->items[$i]->id }})">
+                                                                            <i class="fa fa-pencil"></i>
+                                                                        </button>
+                                                                    @endif
+                                                                    {{-- @if (Auth::user()->hasPermission(['admin.items.delete','contracts.items.delete']) || $contract->dependency_id == Auth::user()->dependency_id) --}}
+                                                                    @if (Auth::user()->hasPermission(['admin.items.delete','contracts.items.delete']))
+                                                                        <button type="button" title="Borrar" class="btn btn-danger btn-icon" onclick="deleteItem({{ $contract->items[$i]->id }})">
+                                                                            <i class="fa fa-trash"></i>
+                                                                        </button>
+                                                                    @endif
                                                                 @endif
-                                                                {{-- @if (Auth::user()->hasPermission(['admin.items.delete','contracts.items.delete']) || $contract->dependency_id == Auth::user()->dependency_id) --}}
-                                                                @if (Auth::user()->hasPermission(['admin.items.delete','contracts.items.delete']))
-                                                                    <button type="button" title="Borrar" class="btn btn-danger btn-icon" onclick="deleteItem({{ $contract->items[$i]->id }})">
-                                                                        <i class="fa fa-trash"></i>
-                                                                    </button>
-                                                                @endif
+
                                                                 @if (Auth::user()->hasPermission(['admin.items.update','contracts.items.update']) || $contract->dependency_id == Auth::user()->dependency_id)
                                                                 {{-- @if (Auth::user()->hasPermission(['admin.items.update','contracts.items.update'])) --}}
                                                                     <button type="button" title="Endosos de Póliza" class="btn btn-primary btn-icon" onclick="itemAwardHistories({{ $contract->items[$i]->id }})">
